+++ Plain/src/main/java/org/example/plain/domain/group/service/GroupServiceImpl.java
@@ -1,9 +1,8 @@
-package org.example.plain.service.impl;
+package org.example.plain.domain.group.service;
 
-import org.example.plain.dto.GroupDTO;
-import org.example.plain.entity.Group;
+import org.example.plain.domain.group.dto.GroupDTO;
+import org.example.plain.domain.group.entity.Group;
 import org.example.plain.repository.GroupRepository;
-import org.example.plain.service.interfaces.GroupService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
