+++ Plain/src/main/java/org/example/plain/domain/groupmember/dto/GroupMemberDTO.java
@@ -1,11 +1,11 @@
-package org.example.plain.dto;
+package org.example.plain.domain.groupmember.dto;
 
 import lombok.Builder;
 import lombok.Data;
-import org.example.plain.entity.Group;
-import org.example.plain.entity.GroupMember;
-import org.example.plain.entity.GroupMemberId;
-import org.example.plain.entity.User;
+import org.example.plain.domain.group.entity.Group;
+import org.example.plain.domain.groupmember.entity.GroupMember;
+import org.example.plain.domain.groupmember.entity.GroupMemberId;
+import org.example.plain.domain.user.entity.User;
 
 @Data
 @Builder
