+++ Plain/src/main/java/org/example/plain/domain/groupmember/service/GroupMemberServiceImpl.java
@@ -1,13 +1,12 @@
-package org.example.plain.service.impl;
+package org.example.plain.domain.groupmember.service;
 
-import org.example.plain.dto.GroupMemberDTO;
-import org.example.plain.entity.Group;
-import org.example.plain.entity.GroupMember;
-import org.example.plain.entity.User;
+import org.example.plain.domain.groupmember.dto.GroupMemberDTO;
+import org.example.plain.domain.group.entity.Group;
+import org.example.plain.domain.groupmember.entity.GroupMember;
+import org.example.plain.domain.user.entity.User;
 import org.example.plain.repository.GroupMemberRepository;
 import org.example.plain.repository.GroupRepository;
 import org.example.plain.repository.UserRepository;
-import org.example.plain.service.interfaces.GroupMemberService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
