+++ Plain/src/main/java/org/example/plain/repository/GroupMemberRepository.java
@@ -1,10 +1,10 @@
 package org.example.plain.repository;
 
 import jakarta.transaction.Transactional;
-import org.example.plain.entity.Group;
-import org.example.plain.entity.GroupMember;
-import org.example.plain.entity.GroupMemberId;
-import org.example.plain.entity.User;
+import org.example.plain.domain.group.entity.Group;
+import org.example.plain.domain.groupmember.entity.GroupMember;
+import org.example.plain.domain.groupmember.entity.GroupMemberId;
+import org.example.plain.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
