+++ Plain/src/main/java/org/example/plain/repository/GroupRepository.java
@@ -1,6 +1,6 @@
 package org.example.plain.repository;
 
-import org.example.plain.entity.Group;
+import org.example.plain.domain.group.entity.Group;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
