+++ Plain/src/main/java/org/example/plain/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package org.example.plain.repository;
 
-import org.example.plain.entity.User;
+import org.example.plain.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
